+++ lib/app/shared/features/currency_withdraw/notifier/withdrawal_amount_notifier/withdrawal_amount_notifier.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:logging/logging.dart';
 import 'package:simple_kit/simple_kit.dart';
@@ -90,7 +91,7 @@ class WithdrawalAmountNotifier extends StateNotifier<WithdrawalAmountState> {
       final baseValue = calculateBaseBalanceWithReader(
         read: read,
         assetSymbol: currency.symbol,
-        assetBalance: double.parse(state.amount),
+        assetBalance: Decimal.parse(state.amount),
       );
 
       _updateBaseConversionValue(
